--- src/widgets/TimeSelector/index.ts
@@ -1,2 +1,0 @@
-export { TimeSelector } from './ui/TimeSelector/TimeSelector';
-export type { Time, DayButtonType } from './model/types';
