+++ src/widgets/time/index.ts
@@ -1,1 +1,4 @@
 export { TimeBottomSheet } from './ui/TimeBottomSheet/TimeBottomSheet';
+
+export { TimeSelector } from './ui/TimeSelector/TimeSelector';
+export type { Time, DayButtonType } from './model/types';
