+++ src/widgets/time/model/consts.ts
