+++ src/widgets/time/model/types.ts
