+++ src/widgets/time/ui/TimeSelector/HourSections/SectionArea/SectionArea.tsx
