+++ src/widgets/time/ui/TimeSelector/HourSections/SectionText/SectionText.style.ts
