+++ src/widgets/time/ui/TimeSelector/TimeDivider/TimeDivider.tsx
