+++ src/widgets/time/ui/TimeSelector/TimeRanges/TimeRange/TimeRange.style.ts
