+++ src/widgets/time/ui/TimeSelector/TimeRanges/TimeRange/TimeRange.tsx
@@ -1,5 +1,5 @@
-import { TIME_COLOR } from '@/widgets/TimeSelector/model/consts';
-import { DragRangeStatus } from '@/widgets/TimeSelector/model/types';
+import { TIME_COLOR } from '@/widgets/time/model/consts';
+import { DragRangeStatus } from '@/widgets/time/model/types';
 
 import { theme } from '@/shared/styles';
