+++ src/widgets/time/ui/TimeSelector/TimeSelector.style.ts
